Feign Client now operates under local and dev
Modification of how each FeignClient is defined. Under the dev profile,
we do not use Zuul; To overcome this, the interface is defined in the
`com.revature.caliber.intercoms.base` package which exposes the same
method as did initially. Since FeignClient supports inheritance, any Feign Client using a Spring
Profile that is !dev && !prod will use the service id's exposed in Zuul,
whereas if the Spring Profile is dev && !prod it will use the Route 53
records